+++ packages/utils/package.json
@@ -1,12 +1,11 @@
 {
-  "name": "@brifui/button",
-  "version": "0.0.1-beta.1",
-  "description": "Button component - Brif UI",
+  "name": "@brifui/utils",
+  "version": "0.0.0",
+  "description": "Brif UI components - utils library",
   "type": "module",
   "sideEffects": false,
   "keywords": [
-    "button",
-    "component",
+    "utils",
     "ui"
   ],
   "files": [
@@ -18,7 +17,6 @@
     "access": "public"
   },
   "scripts": {
-    "typecheck": "tsc --noEmit",
     "build": "tsup src --dts",
     "dev": "tsup src --watch",
     "prepack": "clean-package",
@@ -28,22 +26,28 @@
   "main": "./src/index.ts",
   "module": "./src/index.ts",
   "dependencies": {
-    "@brifui/utils": "workspace:*",
-    "@brifui/styled": "workspace:*",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
+  },
+  "peerDependencies": {
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "@brifui/eslint-config": "workspace:*",
     "@brifui/typescript-config": "workspace:*",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
+    "@microsoft/api-extractor": "^7.49.1",
+    "@rsbuild/plugin-react": "^1.1.0",
+    "@rslib/core": "^0.3.1",
+    "@types/react": "^19.0.5",
+    "@types/react-dom": "^19.0.3",
     "eslint": "^9.15.0",
-    "global-jsdom": "^26.0.0",
-    "jsdom": "^26.0.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "typescript": "5.5.4"
   },
   "tsup": {
+    "minify": false,
     "clean": true,
     "target": "es2019",
     "format": [
