+++ packages/utils/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@brifui/typescript-config/react.json",
+  "include": ["src"],
+  "exclude": ["dist", "build", "node_modules"]
+}
